+++ src/basic/hooks/useCart.ts
@@ -1,0 +1,217 @@
+// TODO: 장바구니 관리 Hook
+// 힌트:
+// 1. 장바구니 상태 관리 (localStorage 연동)
+// 2. 상품 추가/삭제/수량 변경
+// 3. 쿠폰 적용
+// 4. 총액 계산
+// 5. 재고 확인
+//
+// 사용할 모델 함수:
+// - cartModel.addItemToCart
+// - cartModel.removeItemFromCart
+// - cartModel.updateCartItemQuantity
+// - cartModel.calculateCartTotal
+// - cartModel.getRemainingStock
+//
+// 반환할 값:
+// - cart: 장바구니 아이템 배열
+// - selectedCoupon: 선택된 쿠폰
+// - addToCart: 상품 추가 함수
+// - removeFromCart: 상품 제거 함수
+// - updateQuantity: 수량 변경 함수
+// - applyCoupon: 쿠폰 적용 함수
+// - calculateTotal: 총액 계산 함수
+// - getRemainingStock: 재고 확인 함수
+// - clearCart: 장바구니 비우기 함수
+
+import { useCallback, useState } from "react";
+import { CartItem, Coupon, Product } from "../../types";
+// addNotification 타입 지정
+type AddNotificationFn = (
+  message: string,
+  type?: "error" | "success" | "warning"
+) => void;
+
+export function useCart(
+  addNotification: AddNotificationFn,
+  products: Product[]
+) {
+  // TODO: 구현
+  const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
+  const [cart, setCart] = useState<CartItem[]>(() => {
+    const saved = localStorage.getItem("cart");
+    if (saved) {
+      try {
+        return JSON.parse(saved);
+      } catch {
+        return [];
+      }
+    }
+    return [];
+  });
+  const getRemainingStock = (product: Product): number => {
+    const cartItem = cart.find(item => item.product.id === product.id);
+    const remaining = product.stock - (cartItem?.quantity || 0);
+
+    return remaining;
+  };
+
+  const calculateCartTotal = (): {
+    totalBeforeDiscount: number;
+    totalAfterDiscount: number;
+  } => {
+    let totalBeforeDiscount = 0;
+    let totalAfterDiscount = 0;
+
+    cart.forEach(item => {
+      const itemPrice = item.product.price * item.quantity;
+      totalBeforeDiscount += itemPrice;
+      totalAfterDiscount += calculateItemTotal(item);
+    });
+
+    if (selectedCoupon) {
+      if (selectedCoupon.discountType === "amount") {
+        totalAfterDiscount = Math.max(
+          0,
+          totalAfterDiscount - selectedCoupon.discountValue
+        );
+      } else {
+        totalAfterDiscount = Math.round(
+          totalAfterDiscount * (1 - selectedCoupon.discountValue / 100)
+        );
+      }
+    }
+
+    return {
+      totalBeforeDiscount: Math.round(totalBeforeDiscount),
+      totalAfterDiscount: Math.round(totalAfterDiscount),
+    };
+  };
+
+  const addToCart = useCallback(
+    (product: ProductWithUI) => {
+      const remainingStock = getRemainingStock(product);
+      if (remainingStock <= 0) {
+        addNotification("재고가 부족합니다!", "error");
+        return;
+      }
+
+      setCart(prevCart => {
+        const existingItem = prevCart.find(
+          item => item.product.id === product.id
+        );
+
+        if (existingItem) {
+          const newQuantity = existingItem.quantity + 1;
+
+          if (newQuantity > product.stock) {
+            addNotification(
+              `재고는 ${product.stock}개까지만 있습니다.`,
+              "error"
+            );
+            return prevCart;
+          }
+
+          return prevCart.map(item =>
+            item.product.id === product.id
+              ? { ...item, quantity: newQuantity }
+              : item
+          );
+        }
+
+        return [...prevCart, { product, quantity: 1 }];
+      });
+
+      addNotification("장바구니에 담았습니다", "success");
+    },
+    [cart, addNotification, getRemainingStock]
+  );
+
+  const removeFromCart = useCallback((productId: string) => {
+    setCart(prevCart => prevCart.filter(item => item.product.id !== productId));
+  }, []);
+
+  const updateQuantity = useCallback(
+    (productId: string, newQuantity: number) => {
+      if (newQuantity <= 0) {
+        removeFromCart(productId);
+        return;
+      }
+
+      const product = products.find(p => p.id === productId);
+      if (!product) return;
+
+      const maxStock = product.stock;
+      if (newQuantity > maxStock) {
+        addNotification(`재고는 ${maxStock}개까지만 있습니다.`, "error");
+        return;
+      }
+
+      setCart(prevCart =>
+        prevCart.map(item =>
+          item.product.id === productId
+            ? { ...item, quantity: newQuantity }
+            : item
+        )
+      );
+    },
+    [products, removeFromCart, addNotification, getRemainingStock]
+  );
+  const applyCoupon = useCallback(
+    (coupon: Coupon) => {
+      const currentTotal = calculateCartTotal().totalAfterDiscount;
+
+      if (currentTotal < 10000 && coupon.discountType === "percentage") {
+        addNotification(
+          "percentage 쿠폰은 10,000원 이상 구매 시 사용 가능합니다.",
+          "error"
+        );
+        return;
+      }
+
+      setSelectedCoupon(coupon);
+      addNotification("쿠폰이 적용되었습니다.", "success");
+    },
+    [addNotification, calculateCartTotal]
+  );
+
+  const getMaxApplicableDiscount = (item: CartItem): number => {
+    const { discounts } = item.product;
+    const { quantity } = item;
+
+    const baseDiscount = discounts.reduce((maxDiscount, discount) => {
+      return quantity >= discount.quantity && discount.rate > maxDiscount
+        ? discount.rate
+        : maxDiscount;
+    }, 0);
+
+    const hasBulkPurchase = cart.some(cartItem => cartItem.quantity >= 10);
+    if (hasBulkPurchase) {
+      return Math.min(baseDiscount + 0.05, 0.5); // 대량 구매 시 추가 5% 할인
+    }
+
+    return baseDiscount;
+  };
+
+  const calculateItemTotal = (item: CartItem): number => {
+    const { price } = item.product;
+    const { quantity } = item;
+    const discount = getMaxApplicableDiscount(item);
+
+    return Math.round(price * quantity * (1 - discount));
+  };
+
+  return {
+    cart,
+    setCart,
+    addToCart,
+    selectedCoupon,
+    setSelectedCoupon,
+    updateQuantity,
+    applyCoupon,
+    removeFromCart,
+    calculateItemTotal,
+    getRemainingStock,
+    calculateCartTotal,
+  };
+}
